+++ client/components/recipe-form-manual-2.js
@@ -12,18 +12,17 @@ import Collapse from 'react-bootstrap/Collapse'
 import {withRouter} from 'react-router'
 
 class RecipeFormManual extends Component {
-  constructor(props) {
-    super(props)
+  constructor() {
+    super()
     this.state = {
-      id: this.props.currentRecipe.id || 0,
-      name: this.props.currentRecipe.name || '',
-      imgUrl: this.props.currentRecipe.imgUrl || '/recipe-default.jpg',
-      prepTime: this.props.currentRecipe.prepTime || 0,
-      cookTime: this.props.currentRecipe.cookTime || 0,
-      totalTime: this.props.currentRecipe.totalTime || 0,
-      serving: this.props.currentRecipe.serving || 0,
-      steps: this.props.currentRecipe.steps || [],
-      ingredients: this.props.currentRecipe.ingredients || [],
+      name: '',
+      imgUrl: '/recipe-default.jpg',
+      prepTime: 0,
+      cookTime: 0,
+      totalTime: 0,
+      serving: 0,
+      steps: [],
+      ingredients: [],
       validated: false,
       newRecipeId: 0,
       isSaving: false,
@@ -48,7 +47,6 @@ class RecipeFormManual extends Component {
       event.stopPropagation()
     } else {
       const recipe = {
-        id: this.state.id,
         imgUrl: this.state.imgUrl,
         name: this.state.name,
         prepTime: this.state.prepTime,
@@ -58,13 +56,12 @@ class RecipeFormManual extends Component {
         steps: this.state.steps,
         ingredients: this.state.ingredients
       }
+      // if (!this.state.edit)
       this.setState({isSaving: true})
       await dispatchRecipe(recipe)
       // await this.props.addRecipeThunkDispatch(recipe)
-      if (!this.state.edit) {
-        const newRecipeId = this.props.newRecipe.id
-        this.setState({newRecipeId, isSaving: false})
-      }
+      const newRecipeId = this.props.newRecipe.id
+      this.setState({newRecipeId, isSaving: false})
     }
     this.setState({validated: true})
   }
@@ -76,7 +73,7 @@ class RecipeFormManual extends Component {
   }
 
   handleSubmit = e => {
-    console.log('manual state in handlesubmit', this.state)
+    console.log('in manual 2')
     if (!this.state.edit) {
       this.addOrUpdateRecipe(e, this.props.addRecipeThunkDispatch)
     } else {
@@ -92,13 +89,9 @@ class RecipeFormManual extends Component {
       imgUrl,
       name,
       steps,
-      prepTime,
-      cookTime,
-      totalTime,
-      serving,
       ingredients
     } = this.state
-    console.log('state in form manual', this.state)
+
     return (
       <Form onSubmit={this.handleSubmit} noValidate validated={validated}>
         <Row>
@@ -108,7 +101,6 @@ class RecipeFormManual extends Component {
               <Form.Control
                 placeholder="00"
                 id="prepTime"
-                value={prepTime}
                 onChange={this.handleChange}
               />
             </Form.Group>
@@ -118,7 +110,6 @@ class RecipeFormManual extends Component {
               <Form.Control
                 placeholder="00"
                 id="cookTime"
-                value={cookTime}
                 onChange={this.handleChange}
               />
             </Form.Group>
@@ -128,7 +119,6 @@ class RecipeFormManual extends Component {
               <Form.Control
                 placeholder="00"
                 id="totalTime"
-                value={totalTime}
                 onChange={this.handleChange}
               />
             </Form.Group>
@@ -139,7 +129,6 @@ class RecipeFormManual extends Component {
                 type="number"
                 id="serving"
                 placeholder="0"
-                value={serving}
                 required
                 onChange={this.handleChange}
               />
